SolveBugs&Reorganize
- Some bugs solved - Reorganize code
