+++ src/controller/logs.py
@@ -1,0 +1,15 @@
+# -*- coding: utf-8 -*-
+
+import kivy
+kivy.require('1.0.5')
+
+class Logs:
+
+    def login(username):
+        pass
+
+    def logout(username):
+        pass
+
+    def finishList(username):
+        pass
