--- src/interface/barticles.py
@@ -1,12 +1,0 @@
-# -*- coding: utf-8 -*-
-
-import kivy
-kivy.require('1.0.5')
-
-class bArticles:
-
-    def __init__(self):
-        pass
-
-    def updateListArticles(self, articles_list):
-        pass
