--- src/interface/bbuylist.py
@@ -1,21 +1,0 @@
-# -*- coding: utf-8 -*-
-
-import kivy
-kivy.require('1.0.5')
-
-class bBuyList:
-
-    def __init__(self):
-        pass
-
-    def addArticle(self):
-        pass
-
-    def removeArticle(self):
-        pass
-
-    def clearList(self):
-        pass
-
-    def updatePrice(self):
-        pass
