--- src/interface/boptions.py
@@ -1,48 +1,0 @@
-# -*- coding: utf-8 -*-
-
-import kivy
-kivy.require('1.0.5')
-
-class bOptions:
-
-    btstatus_newlist = False
-    btstatus_finishlist = False
-    btstatus_clearlist = False
-    btstatus_next = False
-    btstatus_previous = False
-
-    def __init__(self):
-        pass
-
-    def drawBarOptions(self):
-        pass
-
-    def enableNewListButton(self):
-        pass
-
-    def disableNewListButton(self):
-        pass
-
-    def enableClearListButton(self):
-        pass
-
-    def disableClearListButton(self):
-        pass
-
-    def enableFinishListButton(self):
-        pass
-
-    def disableFinishListButton(self):
-        pass
-
-    def enableNextButton(self):
-        pass
-
-    def disableNextButton(self):
-        pass
-
-    def enablePreviousButton(self):
-        pass
-
-    def disablePreviousButton(self):
-        pass
